Let decimal sequences aggregate themselves
The DecimalSequence wrapper was created to hold helper methods for
a list of integers, so it seems like the best place to put all our
aggregation rules. A sequence will aggregate itself from beginning
to end, subtracting any decreasing terms and adding any increasing
terms.
This method is blind to any validation logic. That is specific to
roman numerals, and should be handled closer to receiving initial
input for the calculator.