Evaluate calculator inputs
The calculator is in charge of sterilizing the input, translating
the roman numerals into decimals, and calculating the result.

This is a pretty primitive version with 2 big issues:
1) the calculation is done using eval, which is dangerous when
accepting user strings. This should be removed in the future.
2) the translation from roman numeral string to integers is not very
Ruby-like. A better array operator could be used here, I just think
it's time for me to move on to other things on this challenge.